New ERT music
The Elite are all about transcendence and living forever and the secrets of the universe and they want to know all this; Some are good, some are bad, some are mixed. But, the good ones don’t ever want to organize, the bad instead are the ones that organize, because they lust after power. Powerful consciences don’t want to dominate other people, they want to empower them, so they don’t tend to get together until things are really late in the game, then they come together. Evil is always defeated, because good is so much stronger. And, we’re on this planet and Einstein’s physics showed it, Maxwell’s physics showed it, all of it, that there is at least twelve dimensions, and now that’s why all the top scientist and billionaires are coming out saying it’s a false hologram, it is artificial. The computers are scanning it and finding tensions points where it is artificially projected and gravity is bleeding in to this universe, that’s what they call dark matter. So, we’re like a thought or a dream that’s like a wisp in some computer program, some god’s mind, whatever. They’re proving it all, it’s all coming out. Now, there’s like this sub transmission zone below the third dimension that’s just turned over the most horrible things, which is what it resonates to, and it’s trying to get up into the third dimension, that’s just a basic level consciousness to launch into the next levels. And our species is already way up to the fifth, sixth dimension consciously, our best people. But there is this big war trying to like, basically destroy humanity, because humanity has free will, and there is a decision to which level we want to go to. We have free will, so evils will have to contend, not just good. And the Elites themselves believe they’re racing, using human technology to try to take our best minds, and build some type of breakaway civilization where they’re going to merge with machines, transcend and break away from the failed species that is man, which is kind of like a false transmission because they’re thinking what they are as ugly and bad, projecting on to themselves instead of believing, no it’s a human test about building us up, and so, Google set up eighteen, nineteen years ago. I knew about this before it was declassified, I’m just saying I have good sources. They want to build a giant artificial system, and Google believes the first artificial intelligence will be a super-computer based on the neuron activities of the hive mind of humanity with billions of people wired into the internet exchange (holy shit), and so all of our thoughts go into it, and we’re actually building a computer that has real neurons in real time that’s also psychically connected to us, that are organic creatures so that they will have current prediction powers, future prediction powers, a true crystal ball. But the big secret is, once you have a crystal ball and know the future, you can add stimuli beforehand and make decisions that can control the future. And so then, it’s the end of consciousness and free will for individuals as we know, and a true two-point-o, in a very bad way, hive mind consciousness with an A.I jacked into everyone, knowing our hopes and dreams, delivering it to us, not in some PKD wire head system, where we plug in and give up our consciousness because of unlimited pleasure, but because we were already wired in and absorbed before we knew it by giving over our consciousness to this system our daily decisions that it was able to manipulate and control into a larger system. There’s now a human counterstrike taking place to shut this up before it gets fully into place and to block these systems, and to try and have an actual debate about where humanity goes, and cut off the paedophiles and psychic vampires that are controlling this A.I system before humanity is destroyed. (((whait a sec. the vamps ctrl AI,how they do it))) The paedophiles, at whatever level, they rule the devil, whatever you want to call, this interdimensional thing, which gives them advanced off world technology, the fallen one, that’s out of this world, is giving them advanced knowledge on how to construct these systems that have already been used before on other populations.( That’s Satan..) That’s Satan. (((explain..))) Satan became something, that you know, the stupid preacher tells you about, who’s totally controlled or something you read about in the news or TV, but this is an interdimensional force that wants to influence us to build something that absorbs us and kills us, rather than the divine free will we are given to build something much better and empowers the species so the species is now making a decision about its entire future. (((Where you getting this from..))) I know from looking at all the data, researching it, studying it, watching the enemy, that’s the big decision that humanity has now got before us.